// event sourcing/kafka_socket_service/src/interfaces/data/i-kafka-keys.ts
export enum WalletKeys {
    PLAYER_ID = "player_id",
    LICENSEE_ID = "licensee_id",
    SKIN_ID = "skin_id"
}

export enum JPEngineKeys {
    PLAYER_ID = "player_id",
    LICENSEE_ID = "licensee_id",
    TRANSACTION_TYPE = "transaction_type",
    JACKPOT_WINS = "jackpot_wins",
    WALLET_APPROVED = "wallet_approved"
}

export enum JackpotBOKeys {
    LICENSEE_ID = "licenseeId",
    PROGRESSIVE_ID = "progressiveInitialParentId",
    PROGRESSIVE_LEVEL_ID = "levelInitialParentId",
    ENGINE_ID = "engineId"
}

export enum UGAPMKeys {
    PLAYER_ID = "player_id",
    LICENSEE_ID = "licensee_id",
    OPERATOR_NAME = "operator",
    SKIN_ID = "skin_id"
}